Created a simple Tasks app
Currently, we have a basic front end deployed through Flask.
It's able to add as many tasks as required, and you can tick the tasks that are done. It strikes out the ones that are done.
Known bug is that clicking the text still strikes out the text but doesn't tick mark the task